% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FittedModel.R
\name{as.character}
\alias{as.character}
\alias{as.character,FittedModel-method}
\title{Convert a \code{\link{FittedModel}} to a String}
\arguments{
\item{x}{the object}
}
\value{
the name of the object
}
\description{
the \code{as.character} implementation for
  \code{\link{FittedModel}}
}
