% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FittedModel.R
\name{FittedModel.as.character}
\alias{FittedModel.as.character}
\title{Convert a \code{FittedModel} to a String}
\usage{
FittedModel.as.character(x)
}
\arguments{
\item{x}{the \code{FittedModel}}
}
\value{
the string
}
\description{
well, convert a \code{FittedModel} to a String
}
