% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.R
\name{xy.order}
\alias{xy.order}
\title{Order Vectors of X and Y Coordinates}
\usage{
xy.order(x, y, x.inc = TRUE, y.inc = FALSE)
}
\arguments{
\item{x}{the first vector}

\item{y}{the second vector}

\item{x.inc}{should the first vector be sorted in increasing order?}

\item{y.inc}{should the second vector be sorted in increasing order?}
}
\value{
a vector with the order
}
\description{
Order a vector \code{x} and \code{y} first based on \code{x} and
  then break ties based on the associated \code{y} values.
}
