% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner.R
\name{regression.applyLearners}
\alias{regression.applyLearners}
\title{Apply a Set if Learners}
\usage{
regression.applyLearners(x, y, learners,
  representations = dataTransformeR::Transformation.applyDefault2D(x, y),
  metric = regressoR.quality::RegressionQualityMetric.default)
}
\description{
Use the learnerSelectoR package to apply a set of learners to a
  set of data representations and pick the approach which generalizes best.
}
